+++ variables/constants.js
@@ -1,6 +1,6 @@
 const chalk = require("chalk");
 
-const VERSION = "0.3.4";
+const VERSION = "0.3.5";
 
 /**
  * Recognized commands that the shell knows.
@@ -219,6 +219,11 @@ const DEFINITIONS = {
     description: "Remove a file or directory from the file system.",
     all: "Remove a file or directory that you specify permanently.",
     usage: ["del <filename>", "del <foldername>"],
+  },
+  tasklist: {
+    description: "Get all running processes on your computer.",
+    all: "WARNING: This command is still in development and subject to changes.\nDisplays the process name, PID, PPID (as well as CMD, CPU, memory, and UID, but these are not available on Windows).\nIf the number to display is <0 or not a number, it is set to 1 and the default size, respectively.",
+    usage: ["tasklist", "tasklist <display>"],
   },
 };